Resizing

Documentation for Resizing.

Resizing.grow_at!

Resizing.grow_beg!

Resizing.grow_end!

Resizing.shrink_at!

Resizing.shrink_beg!

Resizing.shrink_end!

Resizing.unsafe_grow_at!

Resizing.unsafe_grow_beg!

Resizing.unsafe_grow_end!

Resizing.unsafe_shrink_at!

Resizing.unsafe_shrink_beg!

Resizing.unsafe_shrink_end!

Resizing.grow_at! — Method

grow_at!(collection, i::Int, n::Integer) -> Bool

Grow collection by n elements at index i. This does not ensure that new elements are defined. If successful this will return true return true.

Resizing.grow_beg! — Method

grow_beg!(collection, n::Integer) -> Bool

Grow collection by n elements from its first index. This does not ensure that new elements are defined. If successful will return true.

Resizing.grow_end! — Method

grow_end!(collection, n::Integer) -> Bool

Grow collection by n elements from its last index. This does not ensure that new elements are defined. If successful will return true.

Resizing.shrink_at! — Method

shrink_at!(collection, i::Int, n::Integer) -> Bool

Shrink collection by n elements at index i. If successful this will return true.

Resizing.shrink_beg! — Method

shrink_beg!(collection, n::Integer) -> Bool

Deletes n elements from the first index of collection. If successful will return true.

Resizing.shrink_end! — Method

shrink_end!(collection, n::Integer) -> Bool

Deletes n elements from begining at the last index of collection. If successful will return true.

Resizing.unsafe_grow_at! — Method

unsafe_grow_at!(collection, i::Int, n::Integer) -> Nothing

Grows by n elements at index i of collection.

This method assumes that collection will grow without any errors and may result in undefined behavior if the user isn't certain collection can safely grow. For example, an instance of Vector cannot be shared with another instance of Array to change sizes.

Resizing.unsafe_grow_beg! — Method

unsafe_grow_beg!(collection, n) -> Nothing

Grows by n elements at the first index of collection.

This method assumes that collection will grow without any errors and may result in undefined behavior if the user isn't certain collection can safely grow. For example, an instance of Vector cannot be shared with another instance of Array to change sizes.

Resizing.unsafe_grow_end! — Method

unsafe_grow_end!(collection, n) -> Nothing

Grows by n elements at the last index of collection.

This method assumes that collection will grow without any errors and may result in undefined behavior if the user isn't certain collection can safely grow. For example, an instance of Vector cannot be shared with another instance of Array to change sizes.

Resizing.unsafe_shrink_at! — Method

unsafe_shrink_at!(collection, i, n::Integer) -> Nothing

Deletes n elements from index i of collection.

This method assumes that collection will shrink without any errors and may result in undefined behavior if the user isn't certain collection can safely shrink. For example, an instance of Vector cannot be shared with another instance of Array to change sizes. It also assumes that the provided number of elements the will be removed does not exceed the size of collection.

Resizing.unsafe_shrink_beg! — Method

unsafe_shrink_beg!(collection, n::Integer) -> Nothing

Deletes n elements from the first index of collection.

This method assumes that collection will shrink without any errors and may result in undefined behavior if the user isn't certain collection can safely shrink. For example, an instance of Vector cannot be shared with another instance of Array to change sizes. It also assumes that the provided number of elements the will be removed does not exceed the size of collection.

Resizing.unsafe_shrink_end! — Method

unsafe_shrink_end!(collection, n::Integer) -> Nothing

Deletes n elements from the last index of collection.

This method assumes that collection will shrink without any errors and may result in undefined behavior if the user isn't certain collection can safely shrink. For example, an instance of Vector cannot be shared with another instance of Array to change sizes. It also assumes that the provided number of elements the will be removed does not exceed the size of collection.