Resizing

Documentation for Resizing.

Resizing.can_grow_at

Resizing.can_grow_beg

Resizing.can_grow_end

Resizing.can_shrink_at

Resizing.can_shrink_beg

Resizing.can_shrink_end

Resizing.grow_at!

Resizing.grow_beg!

Resizing.grow_end!

Resizing.shrink_at!

Resizing.shrink_beg!

Resizing.shrink_end!

Resizing.unsafe_grow_at!

Resizing.unsafe_grow_beg!

Resizing.unsafe_grow_end!

Resizing.unsafe_shrink_at!

Resizing.unsafe_shrink_beg!

Resizing.unsafe_shrink_end!

Resizing.can_grow_at — Method

can_grow_at(collection, i) -> Bool

Return true if collection is an instance that can grow at index i.

Resizing.can_grow_beg — Method

can_grow_beg(collection) -> Bool

Return true if collection is an instance that can grow from its first index.

Resizing.can_grow_end — Method

can_grow_end(collection) -> Bool

Return true if collection is an instance that can grow from its last index.

Resizing.can_shrink_at — Method

can_shrink_at(collection, i, n) -> Bool

Return true if collection can delete n elements from index, i.

Resizing.can_shrink_beg — Method

can_shrink_beg(collection, n) -> Bool

Return true if collection can delete n elements from its first index.

Resizing.can_shrink_end — Method

can_shrink_end(collection, n) -> Bool

Return true if collection can delete n elements from its last index.

Resizing.grow_at! — Function

grow_at!(collection, i::Int, n::Integer=1) -> Bool

Grow collection by n elements at index i. This does not ensure that new elements are defined. If successful this will return true return true.

Resizing.grow_beg! — Function

grow_beg!(collection, n::Integer=1) -> Bool

Grow collection by n elements from its first index. This does not ensure that new elements are defined. If successful thiw will return true.

Resizing.grow_end! — Function

grow_end!(collection, n::Integer=1) -> Bool

Grow collection by n elements from its last index. This does not ensure that new elements are defined. If successful this will return true.

Resizing.shrink_at! — Function

shrink_at!(collection, i::Int, n::Integer) -> Bool

Shrink collection by n elements at index i. This does not ensure that new elements are defined. If successful this will return true.

Resizing.shrink_beg! — Function

grow_beg!(collection, n::Integer=1) -> Bool

Grow collection by n elements from its first index. This does not ensure that new elements are defined. If successful this will return true.

Resizing.shrink_end! — Function

shrink_end!(collection, n::Integer=1) -> Bool

Deletes n elements from begining at the last index of collection. If successful this will return true.

Resizing.unsafe_grow_at! — Method

unsafe_grow_at!(collection, i, n)

Grows by n elements at index i of collection. This method assumes that can_grow_at(collection, i, n) == true.

Resizing.unsafe_grow_beg! — Method

unsafe_grow_beg!(collection, n)

Grows by n elements at the first index of collection. This method assumes that can_grow_beg(collection, n) == true.

Resizing.unsafe_grow_end! — Method

unsafe_grow_end!(collection, n)

Grows by n elements at the last index of collection. This method assumes that can_grow_end(collection, n) == true.

Resizing.unsafe_shrink_at! — Method

unsafe_shrink_at!(collection, i, n)

Deletes n elements from index i of collection. This method assumes that can_shrink_at(collection, i, n) == true.

Resizing.unsafe_shrink_beg! — Method

unsafe_shrink_beg!(collection, n)

Deletes n elements from the first index of collection. This method assumes that can_shrink_beg(collection, n) == true.

Resizing.unsafe_shrink_end! — Method

unsafe_shrink_end!(collection, n)

Deletes n elements from the last index of collection. This method assumes that can_shrink_end(collection, n) == true.